requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.2
langgraph>=0.2.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
boto3>=1.34.0
pandas>=2.0.0
httpx>=0.25.0
pydantic>=2.0.0
click>=8.0.0
requests>=2.25.0
certifi>=2021.0.0